restructured for github actions, windows/macos/linux builds
