+++ .github/workflows/main.yml
@@ -1,0 +1,44 @@
+# credit: https://data-dive.com/multi-os-deployment-in-cloud-using-pyinstaller-and-github-actions
+name: MPC Autofill Desktop Client
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
+jobs:
+  build:
+    name: Build packages
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        include:
+          - os: macos-latest
+            TARGET: macos
+            CMD_BUILD: cd autofill && pyinstaller autofill.spec
+            OUT_FILE_NAME: autofill
+          - os: windows-latest
+            TARGET: windows
+            CMD_BUILD: cd autofill && pip install pywin32 && pyinstaller autofill.spec
+            OUT_FILE_NAME: autofill.exe
+          - os: ubuntu-latest
+            TARGET: linux
+            CMD_BUILD: cd autofill && pyinstaller autofill.spec
+            OUT_FILE_NAME: autofill
+    steps:
+    - uses: actions/checkout@v1
+    - name: Set up Python 3.9
+      uses: actions/setup-python@v2
+      with:
+        python-version: 3.9
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        cd autofill
+        pip install -r requirements.txt
+    - name: Build with pyinstaller for ${{ matrix.TARGET }}
+      run: ${{matrix.CMD_BUILD}}
+    - name: Upload artifact
+      uses: actions/upload-artifact@v2
+      with:
+        path: ./autofill/dist/${{ matrix.OUT_FILE_NAME }}
+        name: autofill-${{ matrix.TARGET }}
