--- autofill/__init__.py
@@ -1,4 +1,0 @@
-from .constants import *
-from .driver import *
-from .order import *
-from .utils import *
