+++ autofill/requirements.txt
@@ -1,0 +1,8 @@
+attrs
+enlighten
+InquirerPy
+numpy
+pyinstaller
+requests
+selenium
+webdriver-manager
